requirements.txt
selenium==4.19.0
pytest==7.4.3
allure-pytest==2.13.5
allure-python-commons==2.13.5

Faker~=25.3.0
requests~=2.32.2